+++ Sources/MapboxMaps/Foundation/MapboxMap.swift
@@ -181,11 +181,15 @@ public final class MapboxMap: StyleManager {
         __map.destroyRenderer()
     }
 
-    init(map: CoreMap, events: MapEvents) {
+    convenience init(map: CoreMap, events: MapEvents) {
+        self.init(map: map, styleManager: map, events: events)
+    }
+
+    init(map: CoreMap, styleManager: StyleManagerProtocol, events: MapEvents) {
         self.__map = map
         self.events = events
 
-        super.init(with: map, sourceManager: StyleSourceManager(styleManager: map))
+        super.init(with: styleManager, sourceManager: StyleSourceManager(styleManager: map))
 
         __map.createRenderer()
         _isDefaultCameraInitialized.proxied = onCameraChanged.map { _ in true }
@@ -202,17 +206,18 @@ public final class MapboxMap: StyleManager {
 
     // MARK: - Style loading
 
-    /// Loads a `style` from a StyleURI, calling a completion closure when the
-    /// style is fully loaded or there has been an error during load.
+    /// Loads a style from a ``StyleURI``.
     ///
-    /// If style loading started while the other style is already loading, the latter's loading `completion`
-    /// will receive a ``CancelError``. If a style is failed to load, `completion` will receive a ``StyleError``.
+    /// The completion function is called when the style has finished loading. When the style loads, there may be several scenarios:
+    ///  - If the style fails to load due to network issues, several reload attempts will be made, and the callback will be triggered only when the process reaches its terminal state - either successful or failed.
+    /// -  If a new style loading is requested (with a different style URI or JSON) while the previous one is still in progress, the previous one gets cancelled. The previous state receives a ``CancelError`` in the callback.
+    /// -  When the main style document is loaded, the style is considered to have loaded successfully, even if some parts of the style aren't fully loaded.
+    /// -  If the main document fails to load, the callback receives a ``StyleError``.
     ///
     /// - Parameters:
     ///   - styleURI: StyleURI to load
     ///   - transition: Options for the style transition.
-    ///   - completion: Closure called when the style has been fully loaded.
-    ///     If style has failed to load a `MapLoadingError` is provided to the closure.
+    ///   - completion: Closure called when the style has been loaded successfully or failed.
     public func loadStyle(_ styleURI: StyleURI,
                           transition: TransitionOptions? = nil,
                           completion: ((Error?) -> Void)? = nil) {
@@ -233,17 +238,18 @@ public final class MapboxMap: StyleManager {
         loadStyle(styleURI, completion: completion)
     }
 
-    /// Loads a `style` from a JSON string, calling a completion closure when the
-    /// style is fully loaded or there has been an error during load.
+    /// Loads a style from a JSON string.
     ///
-    /// If style loading started while the other style is already loading, the latter's loading `completion`
-    /// will receive a ``CancelError``. If a style is failed to load, `completion` will receive a ``StyleError``.
+    /// The completion function is called when the style has finished loading. When the style loads, there may be several scenarios:
+    ///  - If the style fails to load due to network issues, several reload attempts will be made, and the callback will be triggered only when the process reaches its terminal state - either successful or failed.
+    /// -  If a new style loading is requested (with a different style URI or JSON) while the previous one is still in progress, the previous one gets cancelled. The previous state receives a ``CancelError`` in the callback.
+    /// -  When the main style document is loaded, the style is considered to have loaded successfully, even if some parts of the style aren't fully loaded.
+    /// -  If the main document fails to load, the callback receives a ``StyleError``.
     ///
     /// - Parameters:
     ///   - JSON: Style JSON string
     ///   - transition: Options for the style transition.
-    ///   - completion: Closure called when the style has been fully loaded.
-    ///     If style has failed to load a `MapLoadingError` is provided to the closure.
+    ///   - completion: Closure called when the style has been loaded successfully or failed.
     public func loadStyle(_ JSON: String,
                           transition: TransitionOptions? = nil,
                           completion: ((Error?) -> Void)? = nil) {